테이블 구현 수정
