+++ Backend/src/main/java/com/pathsnap/Backend/ImagePhoto/Entity/ImagePhotoEntity.java
@@ -1,13 +1,19 @@
-package com.pathsnap.Backend.ImagePhoto.domain;
+package com.pathsnap.Backend.ImagePhoto.Entity;
 
-import com.pathsnap.Backend.Image.domain.ImageEntity;
-import com.pathsnap.Backend.PhotoRecord.domain.PhotoRecordEntity;
+import com.pathsnap.Backend.Image.Entity.ImageEntity;
+import com.pathsnap.Backend.PhotoRecord.Entity.PhotoRecordEntity;
 import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Entity
 @Table(name = "ImagePhoto")
 @Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
 public class ImagePhotoEntity {
     @Id
     private String imagePhotoId;
