+++ Backend/src/main/java/com/pathsnap/Backend/PhotoRecord/Entity/PhotoRecordEntity.java
@@ -1,10 +1,12 @@
-package com.pathsnap.Backend.PhotoRecord.domain;
+package com.pathsnap.Backend.PhotoRecord.Entity;
 
-import com.pathsnap.Backend.Image.domain.ImageEntity;
-import com.pathsnap.Backend.ImagePhoto.domain.ImagePhotoEntity;
-import com.pathsnap.Backend.Record.domain.RecordEntity;
+import com.pathsnap.Backend.ImagePhoto.Entity.ImagePhotoEntity;
+import com.pathsnap.Backend.Record.Entity.RecordEntity;
 import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import java.util.Date;
 import java.util.List;
@@ -12,6 +14,9 @@ import java.util.List;
 @Entity
 @Table(name = "PhotoRecord")
 @Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
 public class PhotoRecordEntity {
     @Id
     private String photoRecordId;
@@ -20,7 +25,7 @@ public class PhotoRecordEntity {
     @JoinColumn(name = "record_id")
     private RecordEntity record;
 
-    @OneToMany(mappedBy = "photoRecord")  // 양방향 관계 설정
+    @OneToMany(mappedBy = "photoRecord", cascade = CascadeType.ALL)  // 양방향 관계 설정
     private List<ImagePhotoEntity> imagePhotos;
 
     private int seq;
